Move inventory markup to seperate component
