+++ packages/js/product-editor/src/components/variations-table/inventory-menu-item/index.ts
@@ -1,0 +1,1 @@
+export * from './inventory-menu-item';
